#3 Force propagation of case number
Forced the case number to propagate down recursive directories by including it
as a parameter to all the recursive functions. Need to fix it so it is only optional
on the  function, as all other functions should require it
to be passed from the top level.

The reason the old team's code never  a case number like this new code is
because they unzipped INTO the original case number directory (which is our case
meant they would be unzipping into the NFS server; a big no no). So for this new code
the case number should be queried via  from the path by the top level
code then passed down the functions recursively so it's never lost when unzipping inside
the empty scratch space folder.

Next commit I'll clean up the optional/default parameter situation for case numbers

Towards issue #3