#69 Finished Scan class
Finished implementing the Scan class. Now it represents an
active Scan in progress, with nice methods for the caller to use
when performing a Scan. The caller should check with Scan to see if
a file should be scanned, then the caller scans the file and calls
just_scanned_this_path so Scan can write to the history file if needed.

Still need to add a method for gracefully ending the Scan prematurely.

Towards issue #69