#X Force Test Files to be Old
Fixes a rare case where the test will fail if the repo was downloaded
recently. This is because calling  considers
the mod time of the file when performing the scan. Since the repo was
downloaded recently, the files will be ignored by the scan. This is incorrect
because the test assumes all the files will be considered.

Towards no issue