#133 First Implementation of Platform Extraction
First rough implementation of platform extraction. Works by looking for the user_data
file in the os/etc directory underneath the lumberjack directory. Then searches for a line
that has the text HV_ENV on it. Finally, extracts the value that is set to HV_ENV and maps
that to an internal value used for our search engine.

Towards issue #133